requirements.txt
django
django-cors-headers==3.13.0
djangorestframework==3.14.0
Dj-rest-auth
django-allauth
mysqlclient==2.1.1
Pillow==9.4.0
pytz==2022.7
sqlparse==0.4.3
tzdata==2022.7
